requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
httpx
supabase
